pyproject.toml
[tool.poetry]
name = "furniture"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.2"
python-decouple = "^3.8"
pillow = "^12.1.1"
django-modeltranslation = "^0.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
